+++ test/bigtest/interactors/transactions/TransactionsListInteractor.js
@@ -2,12 +2,15 @@ import {
   collection,
   interactor,
   isPresent,
+  text,
 } from '@bigtest/interactor';
 
 export default interactor(class TransactionsListInteractor {
   static defaultScope = '[data-test-transactions-list]';
 
-  transactions = collection('[class^="mclRow---"][role=row]');
+  transactions = collection('[class^="mclRow---"][role=row]', {
+    rowText: text(),
+  });
 
   isLoaded = isPresent('#transactions-list');
   whenLoaded() {